+++ src/Components/Pages/Profile/orderPage.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const OrderPage = () => {
+  return (
+    <div>
+      order page
+    </div>
+  )
+}
+
+export default OrderPage
